% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInteractionfreeDataset.R
\name{calcInteractionfreeDataset}
\alias{calcInteractionfreeDataset}
\title{Calculate the means of an interaction-free dataset}
\usage{
calcInteractionfreeDataset(means, freq, k.levels)
}
\arguments{
\item{means}{matrix of means, the columns denote the covariates and the rows the different x levels}

\item{freq}{matrix of frequencies, the columns denote the covariates and the rows the different x levels}

\item{k.levels}{numeric vector with the number of levels for all covariates (K1, K2, ...)}
}
\description{
This function calculates an interaction-free dataset (x-effects, K-effects, but no x*K-effects).
}
\keyword{anova}
