% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findConditionalColumns.R
\name{findConditionalColumns}
\alias{findConditionalColumns}
\title{Find the columns based on the fixed levels of the covariates}
\usage{
findConditionalColumns(fixed, k.levels)
}
\arguments{
\item{fixed}{numeric vector with the levels the corresponding covariates should be fixed to (NA for no fixing)}

\item{k.levels}{numeric vector with the number of levels for all covariates (K1, K2, ...)}
}
\description{
This function finds the columns .
}
\keyword{anova}
