% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBalanced.R
\name{isBalanced}
\alias{isBalanced}
\title{Check for balanced data}
\usage{
isBalanced(freq)
}
\arguments{
\item{freq}{matrix of frequencies}
}
\description{
This function tests if frequency matrix is balanced (all frequencies are equal).
}
\examples{
isBalanced(matrix(c(33,27,43,38), 2, 2, T))
isBalanced(matrix(c(30,30,30,30), 2, 2, T))
}
\keyword{anova}
