% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInteractionFree.R
\name{isInteractionFree}
\alias{isInteractionFree}
\title{Check for interaction-free data}
\usage{
isInteractionFree(means, freq, k.levels)
}
\arguments{
\item{freq}{matrix of frequencies}
}
\description{
This function tests if data is interaction-free (x-effects, K-effects, but no x*K-effects).
}
\keyword{anova}
