% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isProportional.R
\name{isProportional}
\alias{isProportional}
\title{Check for proportional data}
\usage{
isProportional(freq)
}
\arguments{
\item{freq}{matrix of frequencies}
}
\description{
This function tests if frequency matrix is proportional (all frequencies are proportional within the x-levels).
}
\examples{
isProportional(matrix(c(33,27,43,38), 2, 2, T))
isProportional(matrix(c(5,10,15,30), 2, 2, T))
}
\keyword{anova}
