% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotunbANOVA.R
\name{plotunbANOVA}
\alias{plotunbANOVA}
\title{Plot the results of unbalancedANOVA.}
\usage{
plotunbANOVA(
  results,
  type,
  showWeights = TRUE,
  showMM = TRUE,
  showEffects = FALSE,
  showIFmeans = FALSE
)
}
\arguments{
\item{results}{object of type 'unbANOVA', result of \link{unbalancedANOVA}}

\item{type}{numeric or character denoting result of which ANOVA type to be ploted (allowed values are any of the following: "I", 1, "II", 2, "III", 3, "ATE")}

\item{showWeights}{boolean, should Weights be added to means?}

\item{showMM}{boolean, should marginal means be plotted?}

\item{showEffects}{boolean, should effects be plotted?}

\item{showIFmeans}{boolean, should interactionfree means be plotted for type 2?}
}
\description{
This function plots the results in nice graphs.
}
\keyword{anova}
